+++ frontend/src/components/dialog/CustomDialog.tsx
@@ -11,7 +11,10 @@ import { toast } from "sonner";
 
 import CreateSessionPrompt from "./Sessions/CreateSession";
 import DeleteSessionButton from "./Sessions/DeleteSessionButton";
-import PlayerTableContainer from "./PlayerContainer";
+
+import { useAllPlayers } from "@/hooks/useAllPlayers";
+import { useGameRecords } from "@/hooks/game/useGameRecords";
+import PlayerTable from "./PlayerTable";
 
 import { useCreateSession } from "@/hooks/sessions/useCreateSession";
 
@@ -34,6 +37,13 @@ export default function CustomDialog({
 }: Props) {
   const { createSession } = useCreateSession();
 
+  const { allPlayers } = useAllPlayers();
+  const { records, updateRecord, addRecord } = useGameRecords(
+    sessionId,
+    allPlayers
+  );
+
+  const playerOptions = allPlayers.map((p) => ({ id: p.id, name: p.name }));
   const handleCreateSession = async () => {
     if (!date) return;
 
@@ -74,7 +84,12 @@ export default function CustomDialog({
           ) : null
         ) : (
           <>
-            <PlayerTableContainer sessionId={sessionId} />
+            <PlayerTable
+              records={records}
+              playerOptions={playerOptions}
+              updateRecord={updateRecord}
+              addRecord={addRecord}
+            />
             <div className="mt-6 flex justify-end">
               <DeleteSessionButton sessionId={sessionId} onDeleted={onClose} />
             </div>
